Spring PW should be ok :)

+++ material/spring-boot-kt-api/src/main/kotlin/com/worldline/training/springbootktapi/service/ProductService.kt
@@ -9,7 +9,7 @@ import org.springframework.stereotype.Service
 
 @Service
 class ProductService(@Autowired val productRepository: ProductRepository) {
-    fun getAll() = productRepository.findAll()
+    fun getAll() = productRepository.findAllByOrderByNameAsc()
 
     // use findByIdOrNull instad of findById because the latter returns an optional<Product> instead of Product?
     fun getById(id: Long) = productRepository.findByIdOrNull(id)
@@ -18,7 +18,11 @@ class ProductService(@Autowired val productRepository: ProductRepository) {
         productRepository.save(product)
     }
 
-    fun deleteOne(id: Long) {
+    fun deleteOne(id: Long): Boolean {
+        if (productRepository.findByIdOrNull(id) == null){
+            return false
+        }
         productRepository.deleteById(id)
+        return true
     }
 }
